backend/requirements.txt
fastapi
uvicorn
requests
motor
pydantic[email]
passlib[bcrypt]
python-dotenv
certifi
